requirements.txt
flask
mercadopago
python-telegram-bot==20.6
requests
python-dotenv
